clean up crew
